+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyChannelPool.java
@@ -174,8 +174,6 @@ public class NettyChannelPool {
     private final Map<NettyChannelGroupKey, NettyChannelGroup> channelGroups = new ConcurrentSkipListMap<>();
     private Map<String, Integer> connectionLimitsPerTag = new ConcurrentSkipListMap<>();
 
-    private int connectionLimit = 1;
-
     private SslContext http2SslContext;
     private SSLContext sslContext;
 
@@ -238,15 +236,6 @@ public class NettyChannelPool {
 
     public boolean isStarted() {
         return this.bootStrap != null;
-    }
-
-    public NettyChannelPool setConnectionLimitPerHost(int limit) {
-        this.connectionLimit = limit;
-        return this;
-    }
-
-    public int getConnectionLimitPerHost() {
-        return this.connectionLimit;
     }
 
     public NettyChannelPool setConnectionLimitPerTag(String tag, int limit) {