+++ xenon-common/src/test/java/com/vmware/xenon/common/TestSubscriptions.java
@@ -319,14 +319,7 @@ public class TestSubscriptions extends BasicTestCase {
         // otherwise we deadlock: the connection for the queued subscribe is used up,
         // no more connections can be created, to that owner.
         this.serviceCount = NettyHttpServiceClient.DEFAULT_CONNECTIONS_PER_HOST / 2;
-        // set the connection limit higher for the test host since it will be issuing parallel
-        // subscribes, POSTs
-        this.host.getClient().setConnectionLimitPerHost(this.serviceCount * 4);
         setUpPeers();
-
-        for (VerificationHost h : this.host.getInProcessHostMap().values()) {
-            h.getClient().setConnectionLimitPerHost(this.serviceCount * 4);
-        }
 
         this.host.waitForReplicatedFactoryServiceAvailable(
                 this.host.getPeerServiceUri(ExampleService.FACTORY_LINK));