+++ xenon-common/src/test/java/com/vmware/xenon/common/http/netty/NettyHttpServiceClientChannelPoolsTest.java
@@ -26,6 +26,7 @@ import org.junit.Test;
 import com.vmware.xenon.common.ServiceClient;
 
 public class NettyHttpServiceClientChannelPoolsTest {
+
     private NettyHttpServiceClient client;
 
     @Before
@@ -53,23 +54,29 @@ public class NettyHttpServiceClientChannelPoolsTest {
     public void testDefaultConnectionLimit() {
         this.client.start();
 
-        assertEquals(NettyHttpServiceClient.DEFAULT_CONNECTIONS_PER_HOST, this.client
-                .getChannelPool()
-                .getConnectionLimitPerHost());
-        assertEquals(NettyHttpServiceClient.DEFAULT_CONNECTIONS_PER_HOST, this.client
-                .getSslChannelPool().getConnectionLimitPerHost());
+        assertEquals(NettyHttpServiceClient.DEFAULT_CONNECTIONS_PER_HOST,
+                this.client.getConnectionLimitPerTag(
+                        NettyHttpServiceClient.CONNECTION_TAG_DEFAULT));
+        assertEquals(NettyHttpServiceClient.DEFAULT_CONNECTION_LIMIT_PER_TAG,
+                this.client.getConnectionLimitPerTag(
+                        NettyHttpServiceClient.CONNECTION_TAG_HTTP2_DEFAULT));
     }
 
     @Test
     public void testSetConnectionLimitBeforeSslChannelPoolStart() {
         int connectionLimit = 11;
 
-        this.client.setConnectionLimitPerHost(connectionLimit);
+        this.client.setConnectionLimitPerTag(NettyHttpServiceClient.CONNECTION_TAG_DEFAULT,
+                connectionLimit);
+        this.client.setConnectionLimitPerTag(NettyHttpServiceClient.CONNECTION_TAG_HTTP2_DEFAULT,
+                connectionLimit);
 
         this.client.start();
 
-        assertEquals(connectionLimit, this.client.getChannelPool().getConnectionLimitPerHost());
-        assertEquals(connectionLimit, this.client.getSslChannelPool().getConnectionLimitPerHost());
+        assertEquals(connectionLimit, this.client.getChannelPool().getConnectionLimitPerTag(
+                NettyHttpServiceClient.CONNECTION_TAG_DEFAULT));
+        assertEquals(connectionLimit, this.client.getSslChannelPool().getConnectionLimitPerTag(
+                NettyHttpServiceClient.CONNECTION_TAG_HTTP2_DEFAULT));
     }
 
     @Test
@@ -78,9 +85,14 @@ public class NettyHttpServiceClientChannelPoolsTest {
 
         this.client.start();
 
-        this.client.setConnectionLimitPerHost(connectionLimit);
+        this.client.setConnectionLimitPerTag(NettyHttpServiceClient.CONNECTION_TAG_DEFAULT,
+                connectionLimit);
+        this.client.setConnectionLimitPerTag(NettyHttpServiceClient.CONNECTION_TAG_HTTP2_DEFAULT,
+                connectionLimit);
 
-        assertEquals(connectionLimit, this.client.getChannelPool().getConnectionLimitPerHost());
-        assertEquals(connectionLimit, this.client.getSslChannelPool().getConnectionLimitPerHost());
+        assertEquals(connectionLimit, this.client.getChannelPool().getConnectionLimitPerTag(
+                NettyHttpServiceClient.CONNECTION_TAG_DEFAULT));
+        assertEquals(connectionLimit, this.client.getSslChannelPool().getConnectionLimitPerTag(
+                NettyHttpServiceClient.CONNECTION_TAG_HTTP2_DEFAULT));
     }
 }